refactor(i18n-provider): convert to TypeScript
