--- src/components/i18n-provider/index.d.ts
@@ -1,1 +1,0 @@
-export { default } from "./i18n-provider";
