--- src/components/i18n-provider/index.js
@@ -1,1 +1,0 @@
-export { default } from "./i18n-provider.component";
