+++ src/components/i18n-provider/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./i18n-provider.component";
+export type { I18nProviderProps } from "./i18n-provider.component";
